sdm: Reset the H/W layers of display
SDM does not reset the H/W layers for the use case, when there are no
App layers present in the draw cycle. It leads to crash in SF dumpsys.
Hence, reset the H/W layers of display in all the use cases. Add SDM
error type which indicates the use case, when there are no App layers
in the draw cycle

CRs-Fixed: 2008789
Change-Id: Ia0c0d237e6e4033993b8d2e46b6a792ff5f18340